MolScribe/requirements.txt
git+https://github.com/thomas0809/MolScribe.git
rdkit
torch
gradio
pillow
pubchempy